перенес LoginView в Controller добавил разделитель в TableView
