+++ Navigation/Detailed screen/Controllers/ProfileViewController.swift
@@ -10,11 +10,13 @@ import UIKit
 final class ProfileViewController: UIViewController {
     
     private var modelStar: [[Any]] = [["Photos"], Modelstar.starArray()]
-    
+
+
     //MARK: - Add Table View
     private lazy var myTableView: UITableView = {
         let tableView = UITableView(frame: .zero, style: .grouped)
-        tableView.separatorStyle = .none
+        tableView.separatorStyle = .singleLine
+        tableView.separatorColor = .black
         tableView.sectionHeaderTopPadding = .zero
         tableView.sectionHeaderHeight = .zero
         tableView.sectionFooterHeight = .zero
@@ -64,9 +66,6 @@ extension ProfileViewController: UITableViewDataSource {
         case 0:
             let cell: PhotosTableViewCell = tableView.dequeueReusableCell(withIdentifier: PhotosTableViewCell.identifier, for: indexPath) as! PhotosTableViewCell
             cell.delegate = self
-            cell.layer.shadowColor = #colorLiteral(red: 0.2588235438, green: 0.7568627596, blue: 0.9686274529, alpha: 1).cgColor
-            cell.layer.shadowOpacity = 0.7
-            cell.layer.shadowOffset = CGSize(width: 4.0, height: 4.0)
             return cell
             
         default:
